Update Stream FAQ.md
Shiny odds were doubled in Gen 6 not Gen 4 

"Generation VI slightly changed how Shiny Pokémon are determined to double the default chances that a Pokémon will be Shiny. The basic formula remains the same, but the threshold for a Pokémon to be Shiny is now 16 instead of 8. This results in a base Shiny probability of approximately 16/65536, or 1/4096." - Shiny wiki

Thank you for ChaoticMeatball for pointing this out!